+++ src/Query/Update.php
@@ -51,6 +51,7 @@ final class Update implements Query
         );
     }
 
+    #[\Override]
     public function parameters(): Sequence
     {
         return $this
@@ -60,6 +61,7 @@ final class Update implements Query
             ->append($this->where->parameters());
     }
 
+    #[\Override]
     public function sql(Driver $driver): string
     {
         /** @var Sequence<string> */
@@ -77,6 +79,7 @@ final class Update implements Query
         );
     }
 
+    #[\Override]
     public function lazy(): bool
     {
         return false;