+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 4.2.0 - 2025-04-09
+
+### Added
+
+- `Formal\Access\Layer\Query\Select::columns()` now accepts row values
+- `Formal\Access\Layer\Query\Insert::into()` now accepts a `Select` query to specify the rows to insert
+
+## 4.1.0 - 2025-03-21
+
+### Added
+
+- Support for `innmind/black-box` `6`
+
+### Fixed
+
+- Support for PHP `8.4`
+
 ## 4.0.0 - 2024-07-14
 
 ### Added